Add check to prevent overwriting envrc
If the file already exists, there's no need to overwrite the envrc. It
is already being replaced with the same exact contents anyway.

+++ ci/setup-environment/seed.sh
@@ -65,10 +65,12 @@ function write_load_balancer_certs() {
 }
 
 function write_envrc() {
-  cat <<- 'EOF' > "${envrc_output_path}"
+  if [ ! -f "${envrc_output_path}" ]; then
+    cat <<- 'EOF' > "${envrc_output_path}"
 eval "$(bbl print-env)"
 export BOSH_DEPLOYMENT=cf
 EOF
+  fi
 }
 
 function commit_capi_ci_private() {
